HDDS-7213 extract ReplicationManagerConfiguration to a separate class
